Rebased Commit

+++ deploy-prod.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
-sudo docker stop not-portfolio
-sudo docker rm not-portfolio
+if sudo docker ps --filter name=not-portfolio | grep not-portfolio; then
+	sudo docker stop not-portfolio
+	sudo docker rm not-portfolio
+fi
 sudo docker compose -f compose.prod.yaml up -d
